+++ frontend-next-migration/src/app/[lng]/(home)/_getPage.ts
@@ -2,7 +2,7 @@ import {MainPageProps} from "@/preparedPages/MainPage";
 import {useServerTranslation} from "@/shared/i18n";
 import {AppExternalLinks} from "@/shared/appLinks/appExternalLinks";
 import {RoutePaths} from "@/shared/appLinks/RoutePaths";
-import {createPage} from "../../_helpers/_createPage";
+import {createPage} from "@/app/_helpers";
 
 export async function _getPage(lng: string) {
     const { t } = await useServerTranslation(lng, "main");
