+++ frontend-next-migration/src/app/[lng]/(home)/page.tsx
@@ -1,7 +1,6 @@
 import {MainPage} from "@/preparedPages/MainPage";
 import {_getPage} from "./_getPage";
-import {createMetadataGenerator} from "@/shared/lib/createMetadataGenerator";
-import {withPageData} from "@/shared/lib/hocs/withPageData";
+import {withPageData,createMetadataGenerator} from "src/app/_helpers";
 
 export const generateMetadata = createMetadataGenerator(_getPage);
