+++ frontend-next-migration/src/app/[lng]/artGame/_getPage.ts
@@ -1,4 +1,4 @@
-import {createPage} from "@/app/_helpers/_createPage";
+import {createPage} from "@/app/_helpers";
 import {GameArtPageProps} from "@/preparedPages/GameArtPage";
 import {makeArtGameSectionsWithI18n} from "@/entities/PresentationPackages";
 import {useServerTranslation} from "@/shared/i18n";
