+++ frontend-next-migration/src/app/[lng]/artGame/page.tsx
@@ -1,7 +1,6 @@
 import { GameArtPage } from '@/preparedPages/GameArtPage';
-import {createMetadataGenerator} from "@/shared/lib/createMetadataGenerator";
+import {withPageData,createMetadataGenerator} from "src/app/_helpers";
 import {_getPage} from "./_getPage";
-import {withPageData} from "@/shared/lib/hocs/withPageData";
 
 export const generateMetadata = createMetadataGenerator(_getPage);
 export default withPageData(GameArtPage, _getPage);
