+++ frontend-next-migration/src/app/[lng]/privacy/_getPage.ts
@@ -1,4 +1,4 @@
-import {createPage} from "@/app/_helpers/_createPage";
+import {createPage} from "@/app/_helpers";
 import {PrivacyPageProps} from "@/preparedPages/PrivacyPage";
 import {makePrivacySectionsWithI18n} from "@/entities/PresentationPackages";
 import {useServerTranslation} from "@/shared/i18n";
