+++ frontend-next-migration/src/app/[lng]/privacy/page.tsx
@@ -1,6 +1,5 @@
 import { PrivacyPage } from '@/preparedPages/PrivacyPage';
-import {createMetadataGenerator} from "@/shared/lib/createMetadataGenerator";
-import {withPageData} from "@/shared/lib/hocs/withPageData";
+import {withPageData,createMetadataGenerator} from "src/app/_helpers";
 import {_getPage} from "./_getPage";
 
 export const generateMetadata = createMetadataGenerator(_getPage);
