+++ frontend-next-migration/src/app/[lng]/team/page.ts
@@ -1,7 +1,6 @@
 import {_getPage} from "./_getPage";
 import {MembersPage} from "@/preparedPages/MembersPage";
-import {withPageData} from "@/shared/lib/hocs/withPageData";
-import {createMetadataGenerator} from "@/shared/lib/createMetadataGenerator";
+import {withPageData,createMetadataGenerator} from "src/app/_helpers";
 
 export const generateMetadata = createMetadataGenerator(_getPage);
 export default withPageData(MembersPage, _getPage);
