+++ frontend-next-migration/src/app/_helpers/index.ts
@@ -1,0 +1,3 @@
+export {createPage} from "./_createPage";
+export {withPageData} from "./withPageData";
+export {createMetadataGenerator} from "./createMetadataGenerator";
