+++ frontend-next-migration/src/app/_helpers/withPageData/index.tsx
